Enable compiling fmpz_mod_mpoly with errors

+++ src/fmpz_mod_mpoly/get_term_var_exp_si.c
@@ -17,7 +17,7 @@ slong fmpz_mod_mpoly_get_term_var_exp_si(const fmpz_mod_mpoly_t A, slong i,
 {
     slong N = mpoly_words_per_exp(A->bits, ctx->minfo);
 
-    if (i >= (ulong) A->length)
+    if (i >= A->length)
         flint_throw(FLINT_ERROR, "fmpz_mod_mpoly_get_term_var_exp_si: index out of range");
 
     return mpoly_get_monomial_var_exp_si(A->exps + N*i, var, A->bits, ctx->minfo);
